requirements.txt
psycopg2-binary==2.9.9
cryptography==43.0.1